collect_money/requirements.txt
celery==5.4.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
redis==5.0.4
